Revamped UI
Revamped UI to have Model, Record, and Train tabs.  Got it to correctly interface with the kinect interface, and to start training.  Need to fix lots of small potential issues in the interface of users doing unexpected things.  Need to fix hardcoded filename and filepaths.  Need to get it working in a build.