Bug fixes and Classify
Fixed a small number of the bugs present.  Added the Classify panel.  Still need to get the kinect interface and file importing working in a Build so that we can then verify that Training works in build (which has been independently verified but not in this project) and figure out where it will create the .nn asset in the built version so we can access that for the Classify process